+++ components/zcash_address/src/kind/unified/fvk.rs
@@ -1,6 +1,6 @@
 use alloc::vec::Vec;
 use core::convert::TryInto;
-use zcash_protocol::constants;
+use zcash_protocol::{address::Revision, constants};
 
 use super::{
     private::{SealedContainer, SealedDataItem},
@@ -81,6 +81,7 @@ impl SealedDataItem for Fvk {
 ///
 /// ```
 /// use zcash_address::unified::{self, Container, Encoding, Item};
+/// use zcash_protocol::address::Revision;
 ///
 /// # #[cfg(not(feature = "std"))]
 /// # fn main() {}
@@ -95,48 +96,71 @@ impl SealedDataItem for Fvk {
 /// let fvks: &[Item<unified::Fvk>] = ufvk.items_as_parsed();
 ///
 /// // And we can create the UFVK from a list of FVKs:
-/// let new_ufvk = unified::Ufvk::try_from_items(fvks.to_vec())?;
+/// let new_ufvk = unified::Ufvk::try_from_items(Revision::R0, fvks.to_vec())?;
 /// assert_eq!(new_ufvk, ufvk);
 /// # Ok(())
 /// # }
 /// ```
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Ufvk(pub(crate) Vec<Item<Fvk>>);
+pub struct Ufvk {
+    pub(crate) revision: Revision,
+    pub(crate) fvks: Vec<Item<Fvk>>,
+}
 
 impl Container for Ufvk {
     type DataItem = Fvk;
 
     fn items_as_parsed(&self) -> &[Item<Fvk>] {
-        &self.0
+        &self.fvks
+    }
+
+    fn revision(&self) -> Revision {
+        self.revision
     }
 }
 
 impl Encoding for Ufvk {}
 
 impl SealedContainer for Ufvk {
-    /// The HRP for a Bech32m-encoded mainnet Unified FVK.
+    /// The HRP for a Bech32m-encoded mainnet Revision 0 Unified FVK.
     ///
     /// Defined in [ZIP 316][zip-0316].
     ///
     /// [zip-0316]: https://zips.z.cash/zip-0316
-    const MAINNET: &'static str = constants::mainnet::HRP_UNIFIED_FVK;
+    const MAINNET_R0: &'static str = constants::mainnet::HRP_UNIFIED_FVK_R0;
 
-    /// The HRP for a Bech32m-encoded testnet Unified FVK.
+    /// The HRP for a Bech32m-encoded testnet Revision 0 Unified FVK.
     ///
     /// Defined in [ZIP 316][zip-0316].
     ///
     /// [zip-0316]: https://zips.z.cash/zip-0316
-    const TESTNET: &'static str = constants::testnet::HRP_UNIFIED_FVK;
+    const TESTNET_R0: &'static str = constants::testnet::HRP_UNIFIED_FVK_R0;
 
     /// The HRP for a Bech32m-encoded regtest Unified FVK.
     ///
     /// Defined in [ZIP 316][zip-0316].
     ///
     /// [zip-0316]: https://zips.z.cash/zip-0316
-    const REGTEST: &'static str = constants::regtest::HRP_UNIFIED_FVK;
+    const REGTEST_R0: &'static str = constants::regtest::HRP_UNIFIED_FVK_R0;
 
-    fn from_inner(fvks: Vec<Item<Self::DataItem>>) -> Self {
-        Self(fvks)
+    /// The HRP for a Bech32m-encoded mainnet Revision 1 Unified FVK.
+    ///
+    /// Defined in [ZIP 316][zip-0316].
+    ///
+    /// [zip-0316]: https://zips.z.cash/zip-0316
+    const MAINNET_R1: &'static str = constants::mainnet::HRP_UNIFIED_FVK_R1;
+    /// The HRP for a Bech32m-encoded testnet Revision 1 Unified FVK.
+    ///
+    /// Defined in [ZIP 316][zip-0316].
+    ///
+    /// [zip-0316]: https://zips.z.cash/zip-0316
+    const TESTNET_R1: &'static str = constants::testnet::HRP_UNIFIED_FVK_R1;
+
+    /// The HRP for a Bech32m-encoded regtest Revision 1 Unified FVK.
+    const REGTEST_R1: &'static str = constants::regtest::HRP_UNIFIED_FVK_R1;
+
+    fn from_inner(revision: Revision, fvks: Vec<Item<Self::DataItem>>) -> Self {
+        Self { revision, fvks }
     }
 }
 
@@ -152,9 +176,9 @@ mod tests {
     use super::{Fvk, ParseError, Ufvk};
     use crate::{
         kind::unified::{private::SealedContainer, Encoding as _},
-        unified::{Item, Typecode},
+        unified::{address::testing::arb_revision, Item, Typecode},
     };
-    use zcash_protocol::consensus::NetworkType;
+    use zcash_protocol::{address::Revision, consensus::NetworkType};
 
     prop_compose! {
         fn uniform128()(a in uniform96(), b in uniform32(0u8..)) -> [u8; 128] {
@@ -207,12 +231,13 @@ mod tests {
 
     prop_compose! {
         fn arb_unified_fvk()(
+            revision in arb_revision(),
             shielded in arb_shielded_fvk(),
             transparent in prop::option::of(arb_transparent_fvk()),
         ) -> Ufvk {
-            let mut items: Vec<_> = transparent.into_iter().chain(shielded).map(Item::Data).collect();
-            items.sort_unstable_by(Item::encoding_order);
-            Ufvk(items)
+            let mut fvks: Vec<_> = transparent.into_iter().chain(shielded).map(Item::Data).collect();
+            fvks.sort_unstable_by(Item::encoding_order);
+            Ufvk { revision, fvks }
         }
     }
 
@@ -244,7 +269,7 @@ mod tests {
             0xdf, 0x63, 0xe7, 0xef, 0x65, 0x6b, 0x18, 0x23, 0xf7, 0x3e, 0x35, 0x7c, 0xf3, 0xc4,
         ];
         assert_eq!(
-            Ufvk::parse_internal(Ufvk::MAINNET, &invalid_padding[..]),
+            Ufvk::parse_internal(Ufvk::MAINNET_R0, &invalid_padding[..]),
             Err(ParseError::InvalidEncoding(
                 "Invalid padding bytes".to_owned()
             ))
@@ -262,7 +287,7 @@ mod tests {
             0x43, 0x8e, 0xc0, 0x3e, 0x9f, 0xf4, 0xf1, 0x80, 0x32, 0xcf, 0x2f, 0x7e, 0x7f, 0x91,
         ];
         assert_eq!(
-            Ufvk::parse_internal(Ufvk::MAINNET, &truncated_padding[..]),
+            Ufvk::parse_internal(Ufvk::MAINNET_R0, &truncated_padding[..]),
             Err(ParseError::InvalidEncoding(
                 "Invalid padding bytes".to_owned()
             ))
@@ -294,7 +319,7 @@ mod tests {
             0x8c, 0x7a, 0xbf, 0x7b, 0x9a, 0xdd, 0xee, 0x18, 0x2c, 0x2d, 0xc2, 0xfc,
         ];
         assert_matches!(
-            Ufvk::parse_internal(Ufvk::MAINNET, &truncated_sapling_data[..]),
+            Ufvk::parse_internal(Ufvk::MAINNET_R0, &truncated_sapling_data[..]),
             Err(ParseError::InvalidEncoding(_))
         );
 
@@ -309,7 +334,7 @@ mod tests {
             0x54, 0xd1, 0x9e, 0xec, 0x8b, 0xef, 0x35, 0xb8, 0x44, 0xdd, 0xab, 0x9a, 0x8d,
         ];
         assert_matches!(
-            Ufvk::parse_internal(Ufvk::MAINNET, &truncated_after_sapling_typecode[..]),
+            Ufvk::parse_internal(Ufvk::MAINNET_R0, &truncated_after_sapling_typecode[..]),
             Err(ParseError::InvalidEncoding(_))
         );
     }
@@ -318,13 +343,16 @@ mod tests {
     fn duplicate_typecode() {
         // Construct and serialize an invalid Ufvk. This must be done using private
         // methods, as the public API does not permit construction of such invalid values.
-        let ufvk = Ufvk(vec![
-            Item::Data(Fvk::Sapling([1; 128])),
-            Item::Data(Fvk::Sapling([2; 128])),
-        ]);
-        let encoded = ufvk.to_jumbled_bytes(Ufvk::MAINNET);
+        let ufvk = Ufvk {
+            revision: Revision::R0,
+            fvks: vec![
+                Item::Data(Fvk::Sapling([1; 128])),
+                Item::Data(Fvk::Sapling([2; 128])),
+            ],
+        };
+        let encoded = ufvk.to_jumbled_bytes(Ufvk::MAINNET_R0);
         assert_eq!(
-            Ufvk::parse_internal(Ufvk::MAINNET, &encoded[..]),
+            Ufvk::parse_internal(Ufvk::MAINNET_R0, &encoded[..]),
             Err(ParseError::DuplicateTypecode(Typecode::SAPLING))
         );
     }
@@ -342,7 +370,7 @@ mod tests {
         ];
 
         assert_eq!(
-            Ufvk::parse_internal(Ufvk::MAINNET, &encoded[..]),
+            Ufvk::parse_internal(Ufvk::MAINNET_R0, &encoded[..]),
             Err(ParseError::OnlyTransparent)
         );
     }