+++ components/zcash_protocol/src/constants/testnet.rs
@@ -51,23 +51,36 @@ pub const B58_SCRIPT_ADDRESS_PREFIX: [u8; 2] = [0x1c, 0xba];
 /// [ZIP 320]: https://zips.z.cash/zip-0320
 pub const HRP_TEX_ADDRESS: &str = "textest";
 
-/// The HRP for a Bech32m-encoded testnet Unified Address.
+/// The HRP for a Bech32m-encoded testnet Revision 0 Unified Address.
 ///
 /// Defined in [ZIP 316][zip-0316].
 ///
 /// [zip-0316]: https://zips.z.cash/zip-0316
-pub const HRP_UNIFIED_ADDRESS: &str = "utest";
+pub const HRP_UNIFIED_ADDRESS_R0: &str = "utest";
 
-/// The HRP for a Bech32m-encoded testnet Unified FVK.
+/// The HRP for a Bech32m-encoded testnet Revision 0 Unified FVK.
 ///
 /// Defined in [ZIP 316][zip-0316].
 ///
 /// [zip-0316]: https://zips.z.cash/zip-0316
-pub const HRP_UNIFIED_FVK: &str = "uviewtest";
+pub const HRP_UNIFIED_FVK_R0: &str = "uviewtest";
 
-/// The HRP for a Bech32m-encoded testnet Unified IVK.
+/// The HRP for a Bech32m-encoded testnet Revision 0 Unified IVK.
 ///
 /// Defined in [ZIP 316][zip-0316].
 ///
 /// [zip-0316]: https://zips.z.cash/zip-0316
-pub const HRP_UNIFIED_IVK: &str = "uivktest";
+pub const HRP_UNIFIED_IVK_R0: &str = "uivktest";
+
+/// The HRP for a Bech32m-encoded regtest Revision 1 Unified Address.
+///
+/// Defined in [ZIP 316][zip-0316].
+///
+/// [zip-0316]: https://zips.z.cash/zip-0316
+pub const HRP_UNIFIED_ADDRESS_R1: &str = "urtest";
+
+/// The HRP for a Bech32m-encoded regtest Revision 1 Unified FVK.
+pub const HRP_UNIFIED_FVK_R1: &str = "urviewtest";
+
+/// The HRP for a Bech32m-encoded regtest Revision 1 Unified IVK.
+pub const HRP_UNIFIED_IVK_R1: &str = "urivktest";