+++ devtools/src/bin/inspect/address.rs
@@ -1,5 +1,5 @@
 use zcash_address::{
-    unified::{self, Encoding},
+    unified::{self, Container as _, Encoding},
     ConversionError, ToAddress, ZcashAddress,
 };
 use zcash_protocol::consensus::NetworkType;
@@ -119,9 +119,10 @@ pub(crate) fn inspect(addr: ZcashAddress) {
                             unified::Receiver::Orchard(data) => {
                                 eprintln!(
                                     "   - Orchard ({})",
-                                    unified::Address::try_from_items(vec![unified::Item::Data(
-                                        unified::Receiver::Orchard(data)
-                                    )])
+                                    unified::Address::try_from_items(
+                                        ua.revision(),
+                                        vec![unified::Item::Data(unified::Receiver::Orchard(data))]
+                                    )
                                     .unwrap()
                                     .encode(&addr.net)
                                 );