+++ zcash_keys/src/keys.rs
@@ -4,7 +4,10 @@ use alloc::vec::Vec;
 use core::fmt::{self, Display};
 
 use zcash_address::unified::{self, Container, Encoding, Item, MetadataItem, Typecode};
-use zcash_protocol::consensus::{self, BlockHeight};
+use zcash_protocol::{
+    address::Revision,
+    consensus::{self, BlockHeight},
+};
 use zip32::{AccountId, DiversifierIndex};
 
 use crate::address::UnifiedAddress;
@@ -962,6 +965,11 @@ impl UnifiedFullViewingKey {
             .chain(self.expiry_time.map(unified::MetadataItem::ExpiryTime));
 
         zcash_address::unified::Ufvk::try_from_items(
+            if self.expiry_height().is_some() || self.expiry_time().is_some() {
+                Revision::R1
+            } else {
+                Revision::R0
+            },
             data_items
                 .map(Item::Data)
                 .chain(meta_items.map(Item::Metadata))
@@ -1291,6 +1299,11 @@ impl UnifiedIncomingViewingKey {
             .chain(self.expiry_time.map(unified::MetadataItem::ExpiryTime));
 
         zcash_address::unified::Uivk::try_from_items(
+            if self.expiry_height.is_some() || self.expiry_time.is_some() {
+                Revision::R1
+            } else {
+                Revision::R0
+            },
             data_items
                 .map(Item::Data)
                 .chain(meta_items.map(Item::Metadata))